Add final services

+++ src/main/java/co/edu/icesi/sgiv/repository/entity/PlanToPlanDetailRepository.java
@@ -3,6 +3,7 @@ package co.edu.icesi.sgiv.repository.entity;
 import co.edu.icesi.sgiv.domain.entity.PlanToPlanDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanToPlanDetailRepository extends JpaRepository<PlanToPlanDetail, Integer> {
+public interface PlanToPlanDetailRepository extends JpaRepository<PlanToPlanDetail, Long> {
+
 
 }
