+++ src/main/java/co/edu/icesi/sgiv/repository/entity/TransportationRepository.java
@@ -3,6 +3,6 @@ package co.edu.icesi.sgiv.repository.entity;
 import co.edu.icesi.sgiv.domain.entity.Transportation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransportationRepository extends JpaRepository<Transportation, Integer> {
+public interface TransportationRepository extends JpaRepository<Transportation, Long> {
 
 }
